Changed parse float to parse int in SignalInfo's getRelativeEfficiency as the value will always be an int as of now.
Added a warning to strings.xml to not mess with the value "na" (for missing/invalid signal readings) if translating to another app and the reason why.

+++ src/com/cc/signalinfo/signals/SignalInfo.java
@@ -164,7 +164,7 @@ public abstract class SignalInfo implements ISignal
         float signalValue =
             AppSetup.DEFAULT_TXT.equals(signals[name])
                 ? -1
-                : Math.abs(Float.parseFloat(signals[name]));
+                : Math.abs(Integer.parseInt(signals[name]));
 
         if (signalValue == -1) {
             return ""; // no value set
